Starting point
Can draw a path using the client's mouse, the render it as a set of epicycles based on the DFT of said path.

User can select the sorting type (default : Amplitude) and enable/disable looping.

The rendering can be replayed, The canvas can be cleared so the user can draw again.